# docs/api/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NAAI ERP API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
...
  "paths": {
...
    "/api/v1/organizations/{organizationId}/commercial-documents": {
      "get": {
        "summary": "List organization-scoped sales invoices, purchase invoices and credit notes"
      },
      "post": {
        "summary": "Create an idempotent exact-amount commercial document draft with customer-validated project and optional contract allocations",
        "requestBody": { "$ref": "#/components/requestBodies/CreateCommercialDocument" }
      }
    },
...
    "/api/v1/organizations/{organizationId}/commercial-documents/relationship-backfill/inventory": {
      "get": {
        "operationId": "inventoryCommercialDocumentRelationshipBackfill",
        "tags": ["commercial-documents", "relationship-backfill"],
        "summary": "Inventory final commercial documents whose canonical project or contract relationships need reviewed correction",
        "parameters": [{ "$ref": "#/components/parameters/OrganizationId" }],
        "responses": {
          "200": {
            "description": "Relationship backfill candidates",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RelationshipBackfillInventoryEnvelope" }
              }
            }
          }
        }
      }
    },
    "/api/v1/organizations/{organizationId}/commercial-documents/{documentId}/relationship-backfill/dry-run": {
      "post": {
        "operationId": "dryRunCommercialDocumentRelationshipBackfill",
        "tags": ["commercial-documents", "relationship-backfill"],
        "summary": "Validate a reverse-replacement relationship correction without mutation",
        "parameters": [
          { "$ref": "#/components/parameters/OrganizationId" },
          { "name": "documentId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "$ref": "#/components/parameters/IfMatch" }
        ],
        "requestBody": { "$ref": "#/components/requestBodies/RelationshipBackfillDryRun" },
        "responses": {
          "200": {
            "description": "Deterministic correction plan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RelationshipBackfillPlanEnvelope" }
              }
            }
          }
        }
      }
    },
    "/api/v1/organizations/{organizationId}/commercial-documents/{documentId}/relationship-backfill/commit": {
      "post": {
        "operationId": "commitCommercialDocumentRelationshipBackfill",
        "tags": ["commercial-documents", "relationship-backfill"],
        "summary": "Commit the exact dry-run plan through audited idempotent reverse replacement",
        "parameters": [
          { "$ref": "#/components/parameters/OrganizationId" },
          { "name": "documentId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "$ref": "#/components/parameters/IfMatch" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": { "$ref": "#/components/requestBodies/RelationshipBackfillCommit" },
        "responses": {
          "200": {
            "description": "Reverse-replacement result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OrganizationDataEnvelope" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/organizations/{organizationId}/expenses/relationship-backfill/inventory": {
      "get": {
        "operationId": "inventoryExpenseRelationshipBackfill",
        "tags": ["expenses", "relationship-backfill"],
        "summary": "Inventory posted expenses whose canonical project or contract relationships need reviewed correction",
        "parameters": [{ "$ref": "#/components/parameters/OrganizationId" }],
        "responses": {
          "200": {
            "description": "Relationship backfill candidates",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RelationshipBackfillInventoryEnvelope" }
              }
            }
          }
        }
      }
    },
    "/api/v1/organizations/{organizationId}/expenses/{expenseId}/relationship-backfill/dry-run": {
      "post": {
        "operationId": "dryRunExpenseRelationshipBackfill",
        "tags": ["expenses", "relationship-backfill"],
        "summary": "Validate an expense reverse-replacement relationship correction without mutation",
        "parameters": [
          { "$ref": "#/components/parameters/OrganizationId" },
          { "name": "expenseId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "$ref": "#/components/parameters/IfMatch" }
        ],
        "requestBody": { "$ref": "#/components/requestBodies/RelationshipBackfillDryRun" },
        "responses": {
          "200": {
            "description": "Deterministic correction plan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RelationshipBackfillPlanEnvelope" }
              }
            }
          }
        }
      }
    },
    "/api/v1/organizations/{organizationId}/expenses/{expenseId}/relationship-backfill/commit": {
      "post": {
        "operationId": "commitExpenseRelationshipBackfill",
        "tags": ["expenses", "relationship-backfill"],
        "summary": "Commit the exact expense dry-run plan through audited idempotent reverse replacement",
        "parameters": [
          { "$ref": "#/components/parameters/OrganizationId" },
          { "name": "expenseId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "$ref": "#/components/parameters/IfMatch" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": { "$ref": "#/components/requestBodies/RelationshipBackfillCommit" },
        "responses": {
          "200": {
            "description": "Reverse-replacement result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OrganizationDataEnvelope" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "OrganizationId": {
        "name": "organizationId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "IfMatch": {
        "name": "If-Match",
        "in": "header",
        "required": true,
        "description": "Current optimistic resource version returned by the record read API.",
        "schema": { "type": "string", "pattern": "^[1-9][0-9]*$" }
      },
...
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1
        }
      },
...
    },
...
    "schemas": {
      "PortableSha256": { "type": "string", "pattern": "^[a-f0-9]{64}$" },
      "RelationshipBackfillCandidate": {
        "type": "object",
        "required": ["id", "version", "state"],
        "properties": {
          "id": { "type": "string" },
          "version": {
            "oneOf": [
              { "type": "string", "pattern": "^[1-9][0-9]*$" },
              { "type": "integer", "minimum": 1 }
            ]
          },
          "state": { "type": "string" },
          "partyId": { "type": "string" },
          "payeePartyId": { "type": "string" },
          "projectIds": { "type": "array", "items": { "type": "string" } },
          "contractIds": { "type": "array", "items": { "type": "string" } }
        },
        "additionalProperties": true
      },
      "RelationshipBackfillInventoryEnvelope": {
        "type": "object",
        "required": ["apiVersion", "requestId", "organizationId", "data"],
        "properties": {
          "apiVersion": { "const": "v1" },
          "requestId": { "type": "string" },
          "organizationId": { "type": "string" },
          "data": {
            "type": "object",
            "required": ["items"],
            "properties": {
              "items": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/RelationshipBackfillCandidate" }
              }
            }
          }
        }
      },
      "RelationshipBackfillPlan": {
        "type": "object",
        "required": ["dryRun", "planHash", "originalId", "originalState", "replacement", "effects"],
        "properties": {
          "dryRun": { "const": true },
          "planHash": { "$ref": "#/components/schemas/PortableSha256" },
          "originalId": { "type": "string" },
          "originalState": { "type": "string" },
          "replacement": { "type": "object", "additionalProperties": true },
          "effects": { "type": "array", "items": { "type": "string" } }
        }
      },
      "RelationshipBackfillPlanEnvelope": {
        "type": "object",
        "required": ["apiVersion", "requestId", "organizationId", "data"],
        "properties": {
          "apiVersion": { "const": "v1" },
          "requestId": { "type": "string" },
          "organizationId": { "type": "string" },
          "data": { "$ref": "#/components/schemas/RelationshipBackfillPlan" }
        }
      },
...
      "OrganizationDataEnvelope": {
        "type": "object",
        "required": ["apiVersion", "requestId", "organizationId", "data"],
        "properties": {
          "apiVersion": { "const": "v1" },
          "requestId": { "type": "string" },
          "organizationId": { "type": "string" },
          "data": {}
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "requestBodies": {
      "RelationshipBackfillDryRun": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["replacement", "reason"],
              "properties": {
                "replacement": { "type": "object", "additionalProperties": true },
                "reason": { "type": "string", "minLength": 1 }
              },
              "additionalProperties": false
            }
          }
        }
      },
      "RelationshipBackfillCommit": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["replacement", "reason", "planHash"],
              "properties": {
                "replacement": { "type": "object", "additionalProperties": true },
                "reason": { "type": "string", "minLength": 1 },
                "planHash": { "$ref": "#/components/schemas/PortableSha256" }
              },
              "additionalProperties": false
            }
          }
        }
      },
...
      "CreateCommercialDocument": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "type",
                "documentNumber",
                "fiscalYear",
                "partyId",
                "documentDate",
                "dueDate",
                "currency",
                "netMinor",
                "taxMinor",
                "grossMinor",
                "controlAccountCode",
                "lines"
              ],
              "properties": {
                "id": { "type": "string", "minLength": 1 },
                "type": {
                  "type": "string",
                  "enum": ["sales_invoice", "purchase_invoice", "credit_note"]
                },
                "documentNumber": { "type": "string", "minLength": 1 },
                "fiscalYear": { "type": "integer", "minimum": 1900, "maximum": 9999 },
                "partyId": { "type": "string", "minLength": 1 },
                "documentDate": { "type": "string", "format": "date" },
                "dueDate": {
                  "type": "string",
                  "format": "date",
                  "description": "Must be on or after documentDate."
                },
                "currency": { "type": "string", "pattern": "^[A-Z]{3}$" },
                "netMinor": { "type": "string", "pattern": "^[0-9]+$" },
                "taxMinor": { "type": "string", "pattern": "^[0-9]+$" },
                "grossMinor": { "type": "string", "pattern": "^[0-9]+$" },
                "controlAccountCode": { "type": "string", "minLength": 1 },
                "migrationSourceExpenseId": {
                  "type": "string",
                  "minLength": 1,
                  "description": "Controlled purchase-invoice migration link to an exact matching legacy expense; the original posted journal must be reversed through the normal journal API before replacement posting."
                },
                "migrationSourceExpenseDate": {
                  "type": "string",
                  "format": "date",
                  "description": "Required together with migrationSourceExpenseId and matched exactly against the legacy expense."
                },
                "lines": {
                  "type": "array",
                  "minItems": 1,
                  "description": "Each line or allocation may carry dimensions.projectId and optional dimensions.contractId. Sales and credit projects must belong to partyId; every contract must belong to its selected project.",
                  "items": { "type": "object" }
                },
                "externalReference": { "type": "object", "additionalProperties": true }
              },
              "additionalProperties": true
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}